requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
altair==5.3.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via httpx
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
blinker==1.8.2
    # via streamlit
cachetools==5.4.0
    # via streamlit
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via streamlit
colorama==0.4.6
    # via
    #   click
    #   tqdm
decorator==4.4.2
    # via moviepy
ffmpeg-python==0.2.0
    # via -r requirements.in
future==1.0.0
    # via ffmpeg-python
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via streamlit
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via replicate
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
imageio==2.34.2
    # via moviepy
imageio-ffmpeg==0.5.1
    # via moviepy
jinja2==3.1.4
    # via
    #   altair
    #   pydeck
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
moviepy==1.0.3
    # via -r requirements.in
numpy==2.0.0
    # via
    #   altair
    #   imageio
    #   moviepy
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
packaging==24.1
    # via
    #   altair
    #   replicate
    #   streamlit
    #   streamlit-ext
pandas==2.2.2
    # via
    #   altair
    #   streamlit
pillow==10.4.0
    # via
    #   imageio
    #   streamlit
proglog==0.1.10
    # via moviepy
protobuf==5.27.2
    # via streamlit
pyarrow==17.0.0
    # via streamlit
pydantic==2.8.2
    # via replicate
pydantic-core==2.20.1
    # via pydantic
pydeck==0.9.1
    # via streamlit
pygments==2.18.0
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.1
    # via pandas
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
replicate==0.28.0
    # via -r requirements.in
requests==2.32.3
    # via
    #   -r requirements.in
    #   moviepy
    #   streamlit
rich==13.7.1
    # via streamlit
rpds-py==0.19.0
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
streamlit==1.36.0
    # via
    #   -r requirements.in
    #   streamlit-ext
streamlit-ext==0.1.10
    # via -r requirements.in
tenacity==8.5.0
    # via streamlit
toml==0.10.2
    # via streamlit
toolz==0.12.1
    # via altair
tornado==6.4.1
    # via streamlit
tqdm==4.66.4
    # via
    #   moviepy
    #   proglog
typing-extensions==4.12.2
    # via
    #   pydantic
    #   pydantic-core
    #   replicate
    #   streamlit
tzdata==2024.1
    # via pandas
urllib3==2.2.2
    # via requests
watchdog==4.0.1
    # via streamlit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
